refactor(client): change file structure of the store
