+++ packages/client/src/components/leaderboardTable/leaderboardTable.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 
 import { useSortableData } from '../../hooks/userSortableData'
-import { PlayerStats } from '../../store/playersStatsSlice'
+import { PlayerStats } from '../../store/reducers/playersStatsSlice'
 
 import styles from './leaderboardTable.module.css'
