+++ packages/client/src/hooks/useAuth.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react'
 import axios, { AxiosError } from 'axios'
-import { UserState } from '../store/userSlice'
+import { UserState } from '../store/reducers/userSlice'
 
 export interface IFormData {
   name: string
