+++ packages/client/src/pages/profile/profile.tsx
@@ -13,7 +13,7 @@ import styles from './profile.module.css'
 
 import { getUser } from '../../store/selectors'
 import { useAppDispatch, useAppSelector } from '../../store/hooks'
-import { editData, logout } from '../../store/userActions'
+import { editData, logout } from '../../store/actions/userActions'
 
 import { store } from '../../store'
