+++ packages/client/src/store/actions/userActions.tsx
@@ -1,8 +1,8 @@
 import { createAsyncThunk } from '@reduxjs/toolkit'
-import UserAPI from '../api/user'
-import AuthAPI from '../api/auth'
+import UserAPI from '../../api/user'
+import AuthAPI from '../../api/auth'
 import { AxiosError, isAxiosError } from 'axios'
-import { axiosErrorHandler } from '../features/api/axiosErrorHandler'
+import { axiosErrorHandler } from '../../features/api/axiosErrorHandler'
 
 type UserDataChangePayload = {
   first_name: string
